worked on the Searchbar and added new Component 	modified:   src/pages/Home/Home.jsx 	new file:   src/pages/Search/FindUsers.jsx
